Feature(MInference): refract KVCache class
